Edit 102 file

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,9 +12,9 @@
 int main(void)
 {
 	int i;
-	double next;
-	double prev;
-	double sum;
+	int next;
+	int prev;
+	int sum;
 
 	prev = 0;
 	i = 0;
